pyproject.toml
[tool.poetry]
name = "flask-app"
version = "0.1.0"
description = "Flask application for n8n integration"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
flask = "2.0.1"
werkzeug = "2.0.3"
google-genai = "^1.9.0"

[tool.poetry.group.dev.dependencies]
yt-dlp = ">=2023.12.30"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
